implemented short type

+++ fuzzer/inputs/header.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
 int main() {
-  int a = 2;
-  printf("hello %d", a);
+  long a = 2;
+  printf("hello %ld", a);
 }
